legacy/requirements.txt

pydantic>=2.7
pandas>=2.1
openpyxl>=3.1
PyMuPDF>=1.24
python-docx>=1.1
pyyaml>=6.0
regex>=2023.12.25

streamlit>=1.35
docx2txt>=0.8
